# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse:
    properties:
      description:
        example: Electronic devices and gadgets
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Electronics
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.CreateCategoryRequest:
    properties:
      description:
        example: Electronic devices and gadgets
        type: string
      name:
        example: Electronics
        type: string
    required:
    - description
    - name
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.GetCategoriesResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse'
        type: array
      total:
        example: 10
        type: integer
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.OrderItemResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      product:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductResponse'
      quantity:
        example: 2
        type: integer
      subtotal:
        example: 199.98
        type: number
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.OrderResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      customer_name:
        example: John Doe
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      order_items:
        items:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.OrderItemResponse'
        type: array
      order_number:
        example: ORD-20260101-001
        type: string
      status:
        example: pending
        type: string
      total_amount:
        example: 199.98
        type: number
      user:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UserResponse'
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        example: Product image
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_primary:
        example: true
        type: boolean
      url:
        example: https://storage.example.com/products/image.jpg
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse'
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Premium wireless headphones with noise cancellation
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      images:
        items:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductImageResponse'
        type: array
      is_active:
        example: true
        type: boolean
      meta:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ResponseMeta'
      name:
        example: Wireless Headphones
        type: string
      price:
        example: 99.99
        type: number
      sku:
        example: WH-1000XM4
        type: string
      stock:
        example: 100
        type: integer
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.ResponseMeta:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.UpdateCategoryRequest:
    properties:
      description:
        example: Electronic devices and gadgets
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Electronics
        type: string
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      last_name:
        example: Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
      role:
        example: user
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_utils_responses.PaginatedResponse:
    properties:
      meta:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.PaginationMeta'
      response:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
    type: object
  github_com_oojoseph67_ecommerce_internal_utils_responses.PaginationMeta:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  github_com_oojoseph67_ecommerce_internal_utils_responses.Response:
    properties:
      data: {}
      error:
        example: invalid request
        type: string
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:1234
info:
  contact:
    email: oojoseph67@gmail.com
    name: Joseph
    url: https://josephbuilds.xyz
  description: A modern e-commerce API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1"
paths:
...
  /categories/:
    get:
      consumes:
      - application/json
      description: Retrieve all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.GetCategoriesResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category (admin only)
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product category (admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing product category (admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /order/:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of the user's orders
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order from the user's cart items
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.OrderResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Create an order
      tags:
      - Orders
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.OrderResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Get an order
      tags:
      - Orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"